vpn: consolidate VPN connection state handling
There used to be two places state was handled: the function that was
called to change the state, and the object method handler for the
VPN connection class.  Since the object method handler was marked
RUN_FIRST in it's g_signal_new() definition, we were destroying
internal class data (like the IPv4 config and IP iface) before other
listeners were able to deal with the state change.

That's all kinda pointless.  Just consolidate the state change
handling and make all the cleanup for the DISCONNECTED/FAILED
states happen after other listeners have had a chance to process
the signal.  It also makes the state change handling a lot clearer.