wifi: various AP mode fixups
Scanning doesn't work well in AP mode, because then you're off-channel
and not serving your clients, which isn't good for anyone.  Plus,
nl80211 refuses to scan in AP mode anyway, so just don't scan.

Second, track the device's mode based on the connection we're activating
or have activated, not based on whatever mode the kernel drivers are
using that second.  That is  more consistent, since there's a race between
when the connection starts being activated, when the device will be in
INFRA mode, and when the supplicant actually gets around to changing the
mode.

Third, fix various other codepaths that weren't quite expecting AP mode.