core: fix crash on device deactivation after 'pending' branch merge
Since the manager now tracks active connections, it needs to hold
a reference to the active connection objects too.  The active
connection teardown code was changed to release that reference
(_active_connection_cleanup()) but the creation code was never
changed to remove the bits that pass ownership over the initial
reference to the NMDevice.  Fix that.