wifi: fix supplicant interface state handling for is_available()
is_available() appears not to have been updated when supplicant
interface states were consolidated between the manager and the
interface objects long ago.  Soon after that, the supplicant
itself added some states.

The wifi device is available to activate whenever the supplicant
is READY or in one of it's operational modes, ie up until
COMPLETED.  Any other state means the supplicant is not ready.