wired: wait for carrier on master devices before proceeding with IP config
If the IP configuration requires slaves/ports, starting IP configuration
before the master device has a carrier (which means at least one slave
or port has a carrier and the device is initialized) is pointless, and
will just take longer or fail.  Recent kernels have cleaned up carrier
handling on bridge/bond devices and the carrier is now a reliable
indication of whether or not the master device is ready to send/receive
traffic.  Thus, if we don't have a carrier, obviously no IP configuration
that depends on connectivity is going to work.